move octokit call to app js
